Add full tests
